Refactor MySqlCommandTests.cs to dynamically cancel tests
MySqlCommandTests.cs contains integration tests that require a database connection.
if the current machine don't have access to a database the test should be ignored
when running in debug mode. Therefore the _unableToRunTests flag was created
for that purpose. If the class Setup() failed to create a mock database that flag
will be set to true. When the flag is 'true', in debug mode the test will be inconclusive
else in 'release' mode the test will fail.